pyproject.toml
[project]
name = "projeto-airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Jean Alves",email = "jean_prog@hotmail.com"}
]
readme = "README.md"
requires-python = "3.11.5"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "apache-airflow (>=3.1.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "dotenv-python (>=0.0.1,<0.0.2)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
